refactor: keep the test-only executor id override with the tests that set it
`raw.rs` held a module that nothing in `raw.rs` sets and nothing outside
the lifecycle tests uses, so it read as part of the task machinery rather
than as scaffolding for one test file. It now lives beside its only
caller, and `thread_id` names it through its full path, which says where
it comes from at the point it is read.

The prose around it oversold Miri and got read the way it was written.
The header said "tests that run under Miri", the only command offered was
a Makefile target that does not exist in this repository, and the comment
at the use site justified the override by what Miri cannot do. Nothing
said these also run on every `cargo test`, which they do: nine of them.

Needing no reactor is what lets Miri run them. It is also why they cost
nothing to run normally. Say both, and give a command that works.
